requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dataclasses-json==0.6.1
dnspython==2.4.2
email-validator==2.1.0.post1
emoji==2.8.0
filelock==3.13.0
filetype==1.2.0
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.0
huggingface-hub==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.322
langdetect==1.0.9
langsmith==0.0.51
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pypdfium2==4.22.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.25
urllib3==2.0.7
Werkzeug==2.3.0
win32-setctime==1.1.0
WTForms==3.1.0
yarl==1.9.2
